Add support for entry of numeric values on Android.
Double-tapping on the control will now allow values to be entered using an
on-screen keyboard on Android (and possibly iOS).

The on-screen keyboard will be of numeric type (a number pad only), as this
is the most appropriate way of entering numbers on a touch-screen device.

Like the earlier changes for touch-enabled devices, this is currently only
tested / validated on Android.