Add functionality for custom formatting and scaling.
This allows the user of the control to display (and parse) fractional values
and / or units (like '%' or 'dB') within the control.

The documentation provides sample code on how to make use of this new
functionality to allow the control to display fractional percentage values.

In addition, a new property for scaling the text inside the control is added.
This is required, because a string like '100.0 %' or '-120 dB' will normally
be too long to fit within the control without font scaling.

This commit closes #5.